requirements.txt
fastapi
uvicorn
python-multipart
# ultralytics
onnxruntime
Pillow
numpy
